chore(analysis): lift classifyDependencies from mcp
Move the pure license classifier (`classifyDependencies`), its supporting
types (`DependencyRef`, `LicenseTier`, `LicenseAuditFlagged`,
`LicenseAuditResult`), and the private `COPYLEFT_PATTERN` regex from
`@opencodehub/mcp/src/tools/license-audit.ts` into a new
`@opencodehub/analysis/src/license-classify.ts`. Re-export from the
analysis barrel.

Why: T-W2-5 (`packages/pack/src/licenses.ts`) needs the same classifier.
`pack` cannot import from `mcp` because that introduces a mcp → pack →
mcp dependency cycle. `analysis` is already a transitive dependency of
both `mcp` and `pack`, so lifting the helper there breaks the cycle
cleanly without adding new package edges.

Mechanical lift only — function body, regex, tier semantics, and
`LicenseAuditResult` shape are byte-identical. The MCP tool now imports
the classifier from `@opencodehub/analysis`; no shim re-export retained.
The mcp-side test (`license-audit.test.ts`) updates only its import
path. A package-local `license-classify.test.ts` mirrors the legacy
9 cases (OK / WARN-on-UNKNOWN / WARN-on-empty / BLOCK-on-GPL /
BLOCK-on-PROPRIETARY / AGPL+SSPL+EUPL+CPAL+OSL+RPL spread / LGPL
non-match / lowercase copyleft / BLOCK-wins-over-WARN).

Refs: T-W2-3 (drift_4 prep, extends spec 005 AC-M5-5).
