+++ packages/mcp/src/tools/license-audit.ts
@@ -25,6 +25,7 @@
 // biome-ignore-all lint/complexity/useLiteralKeys: dot-access disallowed on Record index signatures
 
 import type { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
+import { classifyDependencies, type DependencyRef } from "@opencodehub/analysis";
 import { toolErrorFromUnknown } from "../error-envelope.js";
 import { withNextSteps } from "../next-step-hints.js";
 import { stalenessFromMeta } from "../staleness.js";
@@ -40,76 +41,6 @@ import {
 const LicenseAuditInput = {
   ...repoArgShape,
 };
-
-/**
- * Copyleft license prefix matcher. Upper-cased inputs only — callers must
- * normalise. The regex is anchored so `LGPL-3.0` does NOT match `^GPL`
- * (LGPL is weak copyleft → classified as UNKNOWN/WARN for v1.0, upgraded
- * in a follow-up task).
- */
-const COPYLEFT_PATTERN = /^(GPL|AGPL|SSPL|EUPL|CPAL|OSL|RPL)/;
-
-export interface DependencyRef {
-  readonly id: string;
-  readonly name: string;
-  readonly version: string;
-  readonly ecosystem: string;
-  readonly license: string;
-  readonly lockfileSource: string;
-}
-
-export type LicenseTier = "OK" | "WARN" | "BLOCK";
-
-export interface LicenseAuditFlagged {
-  readonly copyleft: readonly DependencyRef[];
-  readonly unknown: readonly DependencyRef[];
-  readonly proprietary: readonly DependencyRef[];
-}
-
-export interface LicenseAuditResult {
-  readonly tier: LicenseTier;
-  readonly flagged: LicenseAuditFlagged;
-  readonly summary: {
-    readonly total: number;
-    readonly okCount: number;
-    readonly flaggedCount: number;
-  };
-}
-
-/**
- * Pure classification. Exposed so unit tests can assert tier logic
- * without touching the MCP server scaffolding.
- */
-export function classifyDependencies(deps: readonly DependencyRef[]): LicenseAuditResult {
-  const copyleft: DependencyRef[] = [];
-  const unknown: DependencyRef[] = [];
-  const proprietary: DependencyRef[] = [];
-
-  for (const d of deps) {
-    const normalised = d.license.trim().toUpperCase();
-    if (normalised === "" || normalised === "UNKNOWN") {
-      unknown.push(d);
-    } else if (normalised === "PROPRIETARY") {
-      proprietary.push(d);
-    } else if (COPYLEFT_PATTERN.test(normalised)) {
-      copyleft.push(d);
-    }
-  }
-
-  const flaggedCount = copyleft.length + unknown.length + proprietary.length;
-  const hasBlocking = copyleft.length > 0 || proprietary.length > 0;
-  const tier: LicenseTier = hasBlocking ? "BLOCK" : unknown.length > 0 ? "WARN" : "OK";
-
-  return {
-    tier,
-    flagged: { copyleft, unknown, proprietary },
-    summary: {
-      total: deps.length,
-      okCount: deps.length - flaggedCount,
-      flaggedCount,
-    },
-  };
-}
 
 interface LicenseAuditArgs {
   readonly repo?: string | undefined;
